Update readme.md per M. Black
I reorganized the sections and created a semblance of a workflow for someone wanting to connect to YTBNCS or an on-premises web server. The majority of the text is the same -- either rewritten or moved to a new area. I also provided additional contextual information, such as what an API key is (please change/update this section with a better definition -- it's only meant to be a placeholder) or what a state machine is, etc. Furthermore, some of the images appear broken.