Update the Call JavaScript From A WebVI example to use the new JSLI dependencies management feature
The script tags to include Add.js and Multiply.js were moved from the HTML source to the JSLI.
The JSLI was moved to its own library component
The mechanical action of the button to log to the console was changed to reflect the behavior of the code. (i.e. it changes state to indicate that items are being logged continuously until the button is released)