Removing block diagram comments and removing type from JS scripts in the JSLI examples
The two JSLI examples had old block diagram comments telling the user to add script tags to the HTML source. This is no longer necessary so I removed the block diagram comments.
I was also told that the script tags didn't need to include the type information and that the server would know how to handle them, so I removed that from the scripts in the JSLIs to simplify them.